+++ packages/google_maps_flutter/ios/Classes/GoogleMapMarkerController.m
@@ -164,7 +164,27 @@ static UIImage* ExtractIcon(NSObject<FlutterPluginRegistrar>* registrar, NSArray
       image = [UIImage imageNamed:[registrar lookupKeyForAsset:iconData[1]
                                                    fromPackage:iconData[2]]];
     }
+  } else if ([iconData[0] isEqualToString:@"fromBytes"]) {
+    if (iconData.count == 2) {
+      @try {
+        FlutterStandardTypedData* byteData = iconData[1];
+        image = [UIImage imageWithData:[byteData data]];
+      } @catch (NSException* exception) {
+        @throw [NSException exceptionWithName:@"InvalidByteDescriptor"
+                                       reason:@"Unable to interpret bytes as a valid image."
+                                     userInfo:nil];
+      }
+    } else {
+      NSString* error = [NSString
+          stringWithFormat:@"fromBytes should have exactly one argument, the bytes. Got: %lu",
+                           iconData.count];
+      NSException* exception = [NSException exceptionWithName:@"InvalidByteDescriptor"
+                                                       reason:error
+                                                     userInfo:nil];
+      @throw exception;
+    }
   }
+
   return image;
 }
 
@@ -174,12 +194,15 @@ static UIImage* ExtractIcon(NSObject<FlutterPluginRegistrar>* registrar, NSArray
   NSObject<FlutterPluginRegistrar>* _registrar;
   GMSMapView* _mapView;
 }
-- (instancetype)init:(FlutterMethodChannel*)methodChannel mapView:(GMSMapView*)mapView {
+- (instancetype)init:(FlutterMethodChannel*)methodChannel
+             mapView:(GMSMapView*)mapView
+           registrar:(NSObject<FlutterPluginRegistrar>*)registrar {
   self = [super init];
   if (self) {
     _methodChannel = methodChannel;
     _mapView = mapView;
     _markerIdToController = [NSMutableDictionary dictionaryWithCapacity:1];
+    _registrar = registrar;
   }
   return self;
 }