+++ packages/firebase_messaging/CHANGELOG.md
@@ -1,4 +1,8 @@
-## 4.0.1+1
+## 4.0.0+2
+
+* Fix example app's floating action button that stopped working due to a breaking change.
+
+## 4.0.0+1
 
 * Log messages about automatic configuration of the default app are now less confusing.